+++ svs/migrations/0007_partition_sv_tables.py
@@ -114,7 +114,7 @@ operations = [
                 ),
             ),
             ("ensembl_gene_id", models.CharField(max_length=16, null=True)),
-            ("ensembl_transcript_id", models.CharField(max_length=16, null=True)),
+            ("ensembl_transcript_id", models.CharField(max_length=32, null=True)),
             ("ensembl_transcript_coding", models.NullBooleanField()),
             (
                 "ensembl_effect",